Manejo de excepciones en ambas APIs

+++ microservicio_persona/src/main/java/com/backend/microservicio_persona/application/interceptors/PersonaServiceImpl.java
@@ -1,11 +1,16 @@
 package com.backend.microservicio_persona.application.interceptors;
 
+import com.backend.microservicio_persona.application.exceptions.NoChangeDetectedException;
+import com.backend.microservicio_persona.application.exceptions.NoSuchElementException;
+import com.backend.microservicio_persona.application.exceptions.NotEnoughAgeException;
+import com.backend.microservicio_persona.application.exceptions.NotFoundException;
 import com.backend.microservicio_persona.application.repositories.PersonaRepository;
 import com.backend.microservicio_persona.application.services.PersonaService;
 import com.backend.microservicio_persona.domain.entities.Persona;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 public class PersonaServiceImpl implements PersonaService<Persona, UUID> {
@@ -20,21 +25,41 @@ public class PersonaServiceImpl implements PersonaService<Persona, UUID> {
 
     @Override
     public Persona getById(UUID id) {
-        return repository.findById(id).orElse(null);
+        return repository.findById(id).orElseThrow(NotFoundException::new);
     }
 
     @Override
     public Persona create(Persona entity) {
+        if (entity.getEdad() < 18)
+            throw new NotEnoughAgeException();
+
         return repository.save(entity);
     }
 
     @Override
-    public void update(Persona entity) {
+    public void update(UUID id, Persona entity) {
+        entity.setId(id);
+        Optional<Persona> previousEntity = repository.findById(id);
+
+        if (entity.getEdad() < 18)
+            throw new NotEnoughAgeException();
+
+        if (previousEntity.isEmpty())
+            throw new NoSuchElementException();
+
+        if (entity.equals(previousEntity.get()))
+            throw new NoChangeDetectedException();
+
         repository.save(entity);
     }
 
     @Override
     public void delete(UUID id) {
+        Optional<Persona> previousEntity = repository.findById(id);
+
+        if (previousEntity.isEmpty())
+            throw new NoSuchElementException();
+
         repository.deleteById(id);
     }
